+++ tests/testthat/test-html.R
@@ -40,7 +40,7 @@ citeproc_variant <- function() {
   } else if (!rmarkdown::pandoc_available("3.1.8")) {
     # Pandoc 2.14.1 fixed  that
     "new-citeproc-post-2.14.1"
-  } else if (!rmarkdown::pandoc_available("3.8")){
+  } else if (!rmarkdown::pandoc_available("3.8")) {
     "new-citeproc-post-3.1.8"
   } else {
     "new-citeproc-post-3.8"